# perf: Remove string allocations from method Colors.FromArbg